Included the disable_ALE() flag in axisymmetric navier-stokes equations.
Fixed the errors that caused the spherical_navier_stokes tests to fail
under the PARANOID flag. The main error was that the Ellipse geometric object
in src/generic/geom_objects.h was not steady when constructed by passing the
values of the two axes (as opposed to GeometricData). The slight mesh-motion
thus induced caused the wrong answer to be computed in cylindrical coordinates
and the the data in ref_cyl_CR.dat.gz has now been corrected.

Fixed a += error in prescribed_displ_lagr_mult/Makefile.am.

Corrected the comments about the ALE_is_disabled boolean flag so that they
are consistent with what the flag does. This was repeated in a number of 
Equation classes.

Found a few more this-> pointers

Modified the build script to pass the correct cfortran.h flag -DgFortran
or -Df2cFortran for versions 3.0 or 4.0 of the GNU compiler suite based what
the compiler does rather than what it is called! Any other compiler-dependent
flags must be specified in CXXFLAGS and an example is given in the intel
file in the configure_options directory. 



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@15 595e1d82-adf1-4282-bac7-b6073fa2ca8e