Modified the multi-domain boussinesq problems so that the off-diagonal terms are calculated anlytically, which required the addition of one extra reverse lookup function in elements.h that finds the local_eqn_number corresponding to a given global number. In addition, extra functions were provided in navier_stokes and advection-diffusion elements to calculate the derivatives of the velocity and advected field, respectively, with respect to unknowns in the elements.
Added serial multi-domain boussinesq problems to 
multi_physics/boussinesq_convection and added analytic calculation of the
off-diagonal jacobian terms in the refineable single-element driver as well.
Jacobian assembly is about 6 times faster using the analytic methods.

The next stage is to modify the error estimator so that refinement in the
combined element is based on whether the velocity OR temperature error is
above the threshold.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@36 595e1d82-adf1-4282-bac7-b6073fa2ca8e