Fixed all warnings from Intel compiler about partially-overloaded functions and a few over-rides that should have been overloads.
Modified all serial drivers so that the contain no MPI instructions. 
If the library is compiled in parallel, but run in serial then all serial
codes should execute and pass the tests. The only "cheat" here is related to
the HYPRE library which, if compiled in parallel, cannot be run in serial. 
Hence, all serial HYPRE tests are omitted in that case.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@67 595e1d82-adf1-4282-bac7-b6073fa2ca8e