Added a proof-of-concept unstructured adaptive fsi code that should work as soon as Nico's updates to the mesh regeneration are merged. (At present a hack to the triangle_mesh.template.cc is required), so the directory is omitted from the build. This required the fix of a little oversight in the multi domain setup.
Documented static free surface problems in two-d using volume constraints and
prescribed contact angles. Added new drivers that are essentially the same to
solve the same problem using an axisymmetric geometry. Also tidied up the
three-d version of the same code. A little more library code clean-up is 
required, including some global name changes, but this will wait until after
Patrick's "introduction to free surface problems" has been checked in so that
the changes are only done once. 

Probably the most significant change is that the volume constraint sign 
convention has been fixed so that when a fluid volume is enclosed the sign
is positive. This means that for an enclosed air bubble 
(adaptive_bubble_in_channel.cc) the specified volume should be negative.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@223 595e1d82-adf1-4282-bac7-b6073fa2ca8e