Modified the destruction of data in Periodic BoundaryNodes to avoid a segmentation fault. This required the introduction of an additional function because the way that virtual destruction works means that storage allocated in a derived class cannot be deleted in the base class. This seems obvious, but the mechanics got a little involved. The upshot is that if you have a BoundaryNode, then once you are in the Node destructor a dynamic_cast to a BoundaryNode will fail, even if the Node was originally a BoundaryNode before the destruction started ... clear?
In addition, the unstructured adaptive solid and FSI demos have now been
included in the self tests with appropriate test data and the unstructured 3D
adaptive test now uses a global measure so it *should* pass on other machines.

A few tolerance tweaks have also been added to ensure that 
the adaptive drop and bubble codes pass the tests.
 


git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@239 595e1d82-adf1-4282-bac7-b6073fa2ca8e