Corrected the IsotropicStrainEnergyFunction constitutive law so that it (now) correctly computes the second Piola-Kirchhoff stress tensor, rather than the stress tensor used by Green & Zerna (which is almost, but not quite the same). The two are exactly the same if the material is incompressible.
Updated the documentation and validation data for all codes apart from the 
parallel codes because I haven't got MPI installed on this machine. I'll do
the parallel in the next commit.

In addition, for all pseudo-solid mesh applications, the constitutive law
has been changed to GeneralisedHookean.

Finally, added another tweak to the parmetis header to ensure that serial
compilation works.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@247 595e1d82-adf1-4282-bac7-b6073fa2ca8e