Fixed the "Copy of Boundary Node is not a Boundary Node" error. The problem is that HijackedNodes are copies, but they are not BoundaryNodes. This case is now tested and excluded.
Fixed an out-of-range error by commenting out the local 
variable s in temporary_stefan_boltzman.h (which is never used)

For some reason the parallel solver failed to convergene in track_pitch.cc
It worked in serial, but I have fixed this by taking two steps to get to
Re=25, rather than one. I don't know what has changed on the solver side, but
we might want to keep an eye on this.


git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@509 595e1d82-adf1-4282-bac7-b6073fa2ca8e