Hidden the precomputed preconditioner blocks mechanism "under the hood".
* src/generic/block_preconditioner.cc
* src/generic/block_preconditioner.h
modified ndof_types() and and nblock_types so that the returned unsigned
depends on if the precomputed blocks (if passed down). This is required
since the ndof_types or nblock_types may be more fine grain in the
preconditioner above.
This can be overridden with the optional boolean parameter to give the
actual ndof_types or nblock_types.

New function: setup_matrix_vector_product(...), sets up a
MatrixVectorProduct. This is created to "hide" the "under the hood"
machinery of precomputed preconditioner blocks.

* src/navier_stokes/navier_stokes_preconditioners.cc
* src/generic/general_purpose_block_preconditioners.cc
Cleaned up places where "Preconditioner_blocks_have_been_precomputed"
were checked.

git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@539 595e1d82-adf1-4282-bac7-b6073fa2ca8e