A number of changes in the core so that the data used in continuation problems can be stored in the Data objects, rather than at the Problem level. This is achieved by replacing the original TimeStepper(s) by a custom ContinuationStorageScheme, so that all spatial adaptivity can be handled by the same mechanisms as used for history values in time-dependent problems.
The assumption is the continuation and time-stepping will never be done at the
same time (which is true, I think). 

At the moment the new functionality can be activated using the
Problem::Use_continuation_timestepper=true flag, but the default is false. 

This is a mid-way check-in because the next stage is to adapt the bifurcation
detection and continuation routines to use the same idea (auxilliary data 
stored in the place of history values). This is a reasonably large job and so
I want to have a definite check point before going further.

One new interface has been added for arc_length_continuation when the 
continuation parameter is already Data in the problem. 

A few validation tolerance tweaks have also taken place.

I've had to add a new function 
 Mesh::set_mesh_level_time_stepper() that is called to reset the timesteppers
for any Mesh level storage, i.e. SpineData in SpineMeshes and the TimeStepper
that is stored in TriangleMeshes. This should be overloaded if you want to do
continuation problems and you have Mesh level data to worry about.

IMPORTANT: For continuation problems the projection must include all Data not
just those that don't have time-history values (e.g. pressure in 
Navier--Stokes). I have rewritten the projection in the TTaylorHoodElement 
to ensure this works (the storage is already allocates), but not yet for
TCrouzeixRaviart because there is no test case ... yet.


IMPORTANT: The parallel continuation has not been tested for cases with
ExternalHaloData. An error should be thrown, but be aware that this is 
work in progress.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@609 595e1d82-adf1-4282-bac7-b6073fa2ca8e